pyproject.toml
[tool.poetry]
name = "firefly-iii-bot"
version = "0.1.0"
description = ""
authors = ["Denis Malin <denis@malina.page>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
python-dotenv = "^0.19.0"
pydantic = "^1.8.2"
cchardet = "^2.1.7"
aiodns = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
